Testing finished for three methods. Renamed EnforcementAction Folder
