+++ tests/LocalRepositoryTests/EnforcementActions/OrderIdExist.cs
@@ -1,0 +1,49 @@
+﻿using AirWeb.Domain.BaseEntities.Interfaces;
+using AirWeb.Domain.EnforcementEntities.EnforcementActions;
+using AirWeb.LocalRepository.Repositories;
+using AirWeb.TestData.Enforcement;
+using GaEpd.AppLibrary.Domain.Repositories;
+
+
+namespace LocalRepositoryTests.EnforcementActions;
+
+[TestFixture]
+public class OrderIdExistTest
+{
+    private LocalEnforcementActionRepository _repository;
+
+    [SetUp]
+    public void SetUp() => _repository = RepositoryHelper.GetEnforcementActionRepository();
+
+    [TearDown]
+    public void TearDown() => _repository.Dispose();
+
+    [Test]
+    public async Task OrderIdExist_WhenIdExist_ReturnTrue()
+    {
+        // Arrange
+        var existingCO = EnforcementActionData.GetData
+            .OfType<ConsentOrder>()
+            .First();
+
+        // Act
+        var results = await _repository.OrderIdExists(existingCO.OrderId, Guid.NewGuid());
+
+        // Assert
+        results.Should().BeTrue();
+    }
+    [Test]
+    public async Task OrderIdExist_WhenIdDoesNotExist_ReturnFalse()
+    {
+        // Arrange
+        short nonExistingOrderId = 9999;
+
+        // Act
+        var results = await _repository.OrderIdExists(nonExistingOrderId, Guid.NewGuid());
+
+        // Assert
+        results.Should().BeFalse();
+    }
+
+}
+
